余計なコメント取り除いてファイル移した
