+++ src/view/right-sidebar/palette-editor.tsx
@@ -14,6 +14,7 @@ import {
 } from "@/shadcn/components/ui/select";
 import { Slider } from "@/shadcn/components/ui/slider";
 import { getStore, updateStore, useStoreValue } from "@/store/store";
+import { LRUCache } from "@/utils/lru-cache";
 import { useEffect, useRef, useState } from "react";
 
 interface PalettePreviewProps {
@@ -21,34 +22,7 @@ interface PalettePreviewProps {
   pixelLength?: number;
 }
 
-// 汎用LRUキャッシュ（小規模用途）
-class LRUCache<K, V> {
-  private map = new Map<K, V>();
-  constructor(private readonly maxEntries: number) {}
-  get(key: K): V | undefined {
-    const v = this.map.get(key);
-    if (v === undefined) return undefined;
-    // 最近使ったものとして末尾に移動
-    this.map.delete(key);
-    this.map.set(key, v);
-    return v;
-  }
-  set(key: K, value: V): void {
-    if (this.map.has(key)) this.map.delete(key);
-    this.map.set(key, value);
-    if (this.map.size > this.maxEntries) {
-      // 先頭(最も古い)を削除
-      const firstKey = this.map.keys().next().value;
-      if (firstKey !== undefined) this.map.delete(firstKey);
-    }
-  }
-  clear(): void {
-    this.map.clear();
-  }
-}
-
-// プレビュー用 LRU (メモリ節約のため20件程度で十分)
-const palettePreviewCache = new LRUCache<string, ImageData>(20);
+const palettePreviewCache = new LRUCache<string, ImageData>(50);
 
 const PalettePreview = ({ paletteId, pixelLength = 256 }: PalettePreviewProps) => {
   const canvasRef = useRef<HTMLCanvasElement | null>(null);
